Made a much more cleaner code organization. Moved base agent logics to agent and tools folder respectfully and the typed dict to a util file (will move all typed dicts soon).
